#  1. 제목과 본문 사이 빈 칸 한줄 필수 (제목과 본문 분리)
#  2. 생략 가능, 필요 시 자세한 설명 작성 (72자 내외로 마무리)
#  3. 적용한 변경사항과 그것을 만든 이유에 대해 설명
#     - 어떻게보다는 "무엇을", "왜"

+++ src/main/java/com/roommake/channel/controller/ChannelController.java
@@ -4,11 +4,17 @@ import com.roommake.channel.dto.ChannelForm;
 import com.roommake.channel.dto.ChannelInfoDto;
 import com.roommake.channel.service.ChannelService;
 import com.roommake.channel.vo.Channel;
+import com.roommake.resolver.Login;
+import com.roommake.user.security.LoginUser;
+import com.roommake.user.service.UserService;
+import com.roommake.user.vo.User;
+import com.roommake.utils.S3Uploader;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -24,42 +30,53 @@ import java.util.List;
 public class ChannelController {
 
     private final ChannelService channelService;
+    private final UserService userService;
+    private final S3Uploader s3Uploader;
 
     @Operation(summary = "전체 채널 조회", description = "전체 채널정보를 조회한다.")
     @GetMapping("/list")
-    public String list(Model model) {
+    public String list(Model model, Principal principal) {
+        String email = principal != null ? principal.getName() : null;
         List<ChannelInfoDto> channelList = channelService.getAllChannels();
-        List<Channel> participationChannelList = channelService.getChannelsByUserId(1);
+        if (email != null) {
+            User user = userService.getUserByEmail(email);
+            List<Channel> participationChannelList = channelService.getChannelsByUserId(user.getId());
+            model.addAttribute("participationChannelList", participationChannelList);
+        }
         model.addAttribute("channelList", channelList);
-        model.addAttribute("participationChannelList", participationChannelList);
-
         return "channel/list";
     }
 
     @Operation(summary = "채널 등록폼", description = "채널 등록폼를 조회한다.")
     @GetMapping(path = "/create")
+    @PreAuthorize("isAuthenticated()")
     public String createForm(Model model) {
         model.addAttribute("channelForm", new ChannelForm());
 
         return "channel/form";
     }
 
-    @Operation(summary = "채널 등록", description = "채널을 추가한다.")
+    @Operation(summary = "채널 추가", description = "새로운 채널을 추가한다.")
     @PostMapping(path = "/create")
-    public String createChannel(@Valid ChannelForm channelForm, BindingResult errors) {
+    @PreAuthorize("isAuthenticated()")
+    public String createChannel(@Valid ChannelForm channelForm, BindingResult errors, @Login LoginUser loginUser) {
         if (errors.hasErrors()) {
             return "channel/form";
         }
-        channelService.createChannel(channelForm);
+        String imageName = s3Uploader.saveFile(channelForm.getImageFile());
+        channelService.createChannel(channelForm, imageName, loginUser.getId());
 
         return "redirect:/channel/list";
     }
 
+    @Operation(summary = "채널 수정폼", description = "채널 수정폼를 조회한다.")
     @GetMapping(path = "/modify/{channelId}")
-    public String modifyForm(@PathVariable("channelId") int channelId, Principal principal, Model model) {
-        String email = principal != null ? principal.getName() : null;
+    @PreAuthorize("isAuthenticated()")
+    public String modifyForm(@PathVariable("channelId") int channelId, @Login LoginUser loginUser, Model model) {
         Channel channel = channelService.getChannelByChannelId(channelId);
-
+        if (channel.getUser().getId() != loginUser.getId()) {
+            throw new RuntimeException("다른 사용자의 채널은 수정할 수 없습니다.");
+        }
         ChannelForm channelForm = new ChannelForm();
         channelForm.setTitle(channel.getTitle());
         channelForm.setDescription(channel.getDescription());
@@ -74,22 +91,34 @@ public class ChannelController {
         return "channel/form";
     }
 
+    @Operation(summary = "채널 수정", description = "채널 정보를 수정한다.")
     @PostMapping(path = "/modify/{channelId}")
-    public String modifyChannel(@PathVariable("channelId") int channelId, @Valid ChannelForm channelForm, BindingResult errors, Principal principal) {
+    @PreAuthorize("isAuthenticated()")
+    public String modifyChannel(@PathVariable("channelId") int channelId, @Valid ChannelForm channelForm, BindingResult errors, @Login LoginUser loginUser) {
         if (errors.hasErrors()) {
             return "channel/form";
         }
-        String email = principal != null ? principal.getName() : null;
         Channel channel = channelService.getChannelByChannelId(channelId);
-        channelService.modifyChannel(channelForm, channel);
+        if (channel.getUser().getId() != loginUser.getId()) {
+            throw new RuntimeException("다른 사용자의 채널은 수정할 수 없습니다.");
+        }
+        String image = "";
+        if (channelForm.getImageFile() != null) {
+            image = s3Uploader.saveFile(channelForm.getImageFile());
+        }
+        channelService.modifyChannel(channelForm, image, channel);
 
         return "redirect:/channel/post/list/{channelId}";
     }
 
+    @Operation(summary = "채널 삭제", description = "채널을 삭제한다.")
     @GetMapping(path = "/delete/{channelId}")
-    public String deleteChannel(@PathVariable("channelId") int channelId, Principal principal) {
-        String email = principal != null ? principal.getName() : null;
+    @PreAuthorize("isAuthenticated()")
+    public String deleteChannel(@PathVariable("channelId") int channelId, @Login LoginUser loginUser) {
         Channel channel = channelService.getChannelByChannelId(channelId);
+        if (channel.getUser().getId() != loginUser.getId()) {
+            throw new RuntimeException("다른 사용자의 채널은 삭제할 수 없습니다.");
+        }
         channelService.deleteChannel(channel);
 
         return "redirect:/channel/list";
@@ -98,8 +127,9 @@ public class ChannelController {
     @Operation(summary = "채널 참여", description = "채널 참여자를 추가한다.")
     @PostMapping("/addUser")
     @ResponseBody
-    public ResponseEntity<Void> createParticipant(@RequestParam("channelId") int channelId) {
-        channelService.createChannelParticipant(channelId);
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Void> createParticipant(@RequestParam("channelId") int channelId, @Login LoginUser loginUser) {
+        channelService.createChannelParticipant(channelId, loginUser.getId());
 
         return ResponseEntity.ok().build();
     }
@@ -107,8 +137,9 @@ public class ChannelController {
     @Operation(summary = "채널 참여취소", description = "채널 참여자를 삭제한다.")
     @GetMapping("/deleteUser")
     @ResponseBody
-    public ResponseEntity<Void> deleteParticipant(@RequestParam("channelId") int channelId) {
-        channelService.deleteChannelParticipant(channelId);
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Void> deleteParticipant(@RequestParam("channelId") int channelId, @Login LoginUser loginUser) {
+        channelService.deleteChannelParticipant(channelId, loginUser.getId());
 
         return ResponseEntity.ok().build();
     }
